moved some of the logic from main.tsx to a separate file
